Screen

+++ main.py
@@ -1,16 +1,26 @@
-from screenshoter import ScreenShoter, ScreenFromTime
-from hotkeyer import KeyListener
-import os, time
+from tools.screenshoter import ScreenShoter, ScreenFromTime
+from tools.hotkeyer import KeyListener
+import os, time, configparser
 
-from_time = ScreenFromTime(ScreenShoter("time", "main/"), 3)
-from_key = KeyListener(ScreenShoter("key", "key/"))
-for i in range(3):
+os.mkdir("cache")
+config = configparser.ConfigParser()
+config.read("config.ini")
+pause = int(config["Setting"]["pause"])
+from_time = ScreenFromTime(ScreenShoter("time", "cache/main/"), pause)
+from_key = KeyListener(ScreenShoter("key", "cache/key/"))
+to_remove = []
+
+
+
+for i in range(int(config["Setting"]["area_num"])):
     from_key.screen_shoter.add_area()
-    img = ScreenShoter.queue.get()
-    img[0].save(img[1])
+    if i == 0:
+        img = ScreenShoter.queue.get()
+        img[0].save(img[1])
+        to_remove.append(img)
     from_key.screen_shoter.fix_coord()
-    to_remove = img
-    os.remove(img[1])
+img = to_remove.pop()
+os.remove(img[1])
 
 from_key.start()
 from_time.start()
@@ -18,7 +28,8 @@ from_time.start()
 
 
 def send(img):
-    print("Sending...")
+    print(f"{img[1]} sending...")
+    # img[0].save(img[1])
     time.sleep(0.1)
     print("Sended")
 
@@ -26,21 +37,33 @@ try:
     while True:
         img = ScreenShoter.queue.get()
 
-        ishotkey = img[1].startswith("key")
-        for cached_img in range(100):
-            notequal = img[0].tobytes() != to_remove[0].tobytes()
+        ishotkey = "key" in img[1]
+        if ishotkey:
+            send(img)
+            continue
 
-        if ishotkey or notequal:
+        notequal = True
+        for i, cached_img in enumerate(to_remove):
+            if img[2] == cached_img[2]:
+                notequal = img[0].tobytes() != cached_img[0].tobytes()
+                to_remove[i] = img
+                break
+        else:
+            to_remove.append(img)
+
+        if notequal:
             send(img)
         else:
             print("Not send")
 
-        to_remove = img
-
-except:
+except :
+    os.chdir("cache")
     dirs = os.listdir()
     for dir in dirs:
-        if os.path.isdir(dir) and dir[0].isalpha():
+        if os.path.isdir(dir) and dir[0].isalpha() and dir != "tools":
             os.removedirs(dir)
+
+    os.chdir("../")
+    os.removedirs("cache")
